Split accounts into files
